backend/requirements/base.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
celery[redis]==5.3.4
django-celery-beat==2.5.0
channels==4.0.0
channels-redis==4.2.0
Pillow==10.1.0
python-decouple==3.8
cryptography>=41.0.0
dj-database-url==2.1.0